[BED-25] Relocate to top-package 'shop.item'
Took 11 minutes
