+++ src/main/java/net/alphalightning/bedwars/game/ui/shop/item/FastBuyGui.java
@@ -1,9 +1,8 @@
-package net.alphalightning.bedwars.game.ui;
+package net.alphalightning.bedwars.game.ui.shop.item;
 
-import net.alphalightning.bedwars.game.ui.item.BuyableItem;
-import net.alphalightning.bedwars.game.ui.item.BuyablePotionItem;
-import net.alphalightning.bedwars.game.ui.item.CurrentItem;
-import net.alphalightning.bedwars.game.ui.legacy.*;
+import net.alphalightning.bedwars.game.ui.shop.item.items.BuyableItem;
+import net.alphalightning.bedwars.game.ui.shop.item.items.BuyablePotionItem;
+import net.alphalightning.bedwars.game.ui.shop.item.items.CurrentItem;
 import org.bukkit.Material;
 import org.bukkit.potion.PotionEffectType;
 import org.bukkit.potion.PotionType;
