+++ src/main/java/net/alphalightning/bedwars/game/ui/shop/item/legacy/KitsItem.java
@@ -1,4 +1,4 @@
-package net.alphalightning.bedwars.game.ui.legacy;
+package net.alphalightning.bedwars.game.ui.shop.item.legacy;
 
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.translation.GlobalTranslator;
